new acl lists

+++ www/install/updatescripts/functions.inc.php
@@ -1,21 +1,35 @@
 <?php
-function apply_write_acl($acl_id, $old_acl_write){
+function apply_write_acl($acl_id, $old_acl_write=0, $delete=true){
 	global $GO_SECURITY;
 	
 	$selectdb = new db();
 	$updatedb = new db();
-
 	
 
 	$sql = "UPDATE go_acl SET level=1 WHERE acl_id=".$acl_id." AND level=-1";
 	$updatedb->query($sql);
 
-	$selectdb->query("SELECT * FROM go_acl WHERE acl_id=".$old_acl_write);
-	while($record = $selectdb->next_record()){
-		$sql = "UPDATE go_acl SET level=3 WHERE acl_id=".$acl_id." AND user_id=".$record['user_id']." AND group_id=".$record['group_id'];
-		$updatedb->query($sql);
+	if($old_acl_write>0){
+		$selectdb->query("SELECT * FROM go_acl WHERE acl_id=".$old_acl_write);
+		while($record = $selectdb->next_record()){
+			$r['acl_id']=$acl_id;
+			$r['user_id']=$record['user_id'];
+			$r['group_id']=$record['group_id'];
+			$r['level']=3;
+
+			$updatedb->replace_row('go_acl', $r);
+		}
+		if($delete)
+			$GO_SECURITY->delete_acl($old_acl_write);
 	}
 
-	$GO_SECURITY->delete_acl($old_acl_write);
+	$sql = "SELECT * FROM go_acl_items WHERE id=".$acl_id;
+	$selectdb->query($sql);
+	$acl = $selectdb->next_record();
+
+	if(!empty($acl['user_id']))
+		$GO_SECURITY->add_user_to_acl($acl['user_id'], $acl['id'], 4);
+
+	$GO_SECURITY->add_group_to_acl(1, $acl_id, 4);
 
 }
